fix: swap -e and +e in helper function
In a helper function we turned off exiting on error, then enabled it
after calling an external utility. This had the behavior of forcing
later calls in the script to immediately exit whenever a non-zero error
code was returned by anything. Unfortunatly, we were testing the result
of a helper function (`CheckSig`) by testing `$?`. If it was non-zero,
we'd immediately exit the program, leading to a failure to actually
generate signatures, and a failure to finish installation.

Change the code to set the flags in the correct order.